retriever/requirements.txt
pdfplumber
faiss-cpu
requests
flask
tqdm
numpy
python-dotenv